aggregator/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
elasticsearch==7.13.1
psycopg2-binary==2.9.1
# psycopg2==2.9.1
redis==3.5.3
# celery==5.1.2
rq==1.10.1
apscheduler==3.7.0
opencensus-ext-flask==0.7.3
opencensus-ext-azure==1.0.7
prometheus-flask-exporter==0.18.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
tzdata==2024.1
flask-cors==3.0.10
gunicorn==20.1.0


# Agents

celery==5.1.2
google-cloud-speech==2.5.1
google-cloud-language==2.3.0
google-cloud-vision==2.3.1
google-cloud-videointelligence==2.3.1
opencv-python-headless==4.5.3.56
pydub==0.25.1
numpy==1.21.0
scipy==1.7.0
noisereduce==2.0.0
spacy==3.7.5
textblob==0.15.3
transformers==4.8.2
torch==1.9.0
redis==3.5.3
requests==2.26.0
openai==1.40.4

